requirements.txt
# Core deep learning
torch==1.10.1+cu111
torchvision==0.11.2+cu111
torchaudio==0.10.1+cu111
triton==3.0.0

# MMSegmentation stack
mmcv-full==1.7.1
mmsegmentation==0.30.0
mmcls==0.25.0
mmengine==0.10.7

# Vision / image processing
opencv-python==4.11.0.86
pillow==11.3.0
scikit-image==0.19.3
ImageIO==2.37.2
PyWavelets==1.4.1
tifffile==2024.8.30
timm==0.5.4

# Scientific computing
numpy==1.21.0
scipy==1.10.1
matplotlib==3.8.4

# Utilities
einops==0.4.1
fvcore==0.1.5.post20221221
iopath==0.1.10
addict==2.4.0
easydict==1.13
yacs==0.1.8
yapf==0.33.0
prettytable==3.16.0
tabulate==0.9.0
termcolor==3.1.0
tqdm==4.67.3
PyYAML==6.0.3
packaging==26.0

# Experiment tracking
wandb==0.25.1

# Misc
GitPython==3.1.46
requests==2.32.5
six==1.17.0
protobuf==6.33.6
portalocker==3.2.0
sympy==1.14.0
